requirements.txt
scikit-learn
pandas
numpy
pyyaml
joblib
